KTO-604: Render external HTML using StyledSectionHTML - Refactor ePeruste kuvaus and osaamisalat use a shared component for rendering external HTML with styles
